requirements.txt
opencv-python
requests
google
google-cloud
google-cloud-vision
yt-dlp
youtube_dl
pandas
nltk
python-dotenv
google-cloud-speech
google-cloud-storage
audio_metadata